+++ psutil/_psposix.py
@@ -121,7 +121,12 @@ def wait_pid(pid, timeout=None):
 
 
 def disk_usage(path):
-    """Return disk usage associated with path."""
+    """Return disk usage associated with path.
+    Note: UNIX usually reserves 5% disk space which is not accessible
+    by user. In this function "total" and "used" values reflect the
+    total and used disk space whereas "free" and "percent" represent
+    the "free" and "used percent" user disk space.
+    """
     try:
         st = os.statvfs(path)
     except UnicodeEncodeError:
@@ -137,14 +142,29 @@ def disk_usage(path):
             st = os.statvfs(path)
         else:
             raise
-    free = (st.f_bavail * st.f_frsize)
+
+    # Total space which is only available to root (unless changed
+    # at system level).
     total = (st.f_blocks * st.f_frsize)
-    used = (st.f_blocks - st.f_bfree) * st.f_frsize
-    percent = usage_percent(used, total, _round=1)
+    # Remaining free space usable by root.
+    avail_to_root = (st.f_bfree * st.f_frsize)
+    # Remaining free space usable by user.
+    avail_to_user = (st.f_bavail * st.f_frsize)
+    # Total space being used in general.
+    used = (total - avail_to_root)
+    # Total space which is available to user (same as 'total' but
+    # for the user).
+    total_user = used + avail_to_user
+    # User usage percent compared to the total amount of space
+    # the user can use. This number would be higher if compared
+    # to root's because the user has less space (usually -5%).
+    usage_percent_user = usage_percent(used, total_user, _round=1)
+
     # NB: the percentage is -5% than what shown by df due to
     # reserved blocks that we are currently not considering:
     # https://github.com/giampaolo/psutil/issues/829#issuecomment-223750462
-    return sdiskusage(total, used, free, percent)
+    return sdiskusage(
+        total=total, used=used, free=avail_to_user, percent=usage_percent_user)
 
 
 @memoize